+++ fastmri_recon/data/sequences/oasis_sequences.py
@@ -133,7 +133,7 @@ class Masked2DSequence(Oasis2DSequence):
         k_shape = images[0].shape
         kspaces = np.empty_like(images, dtype=np.complex64)
         mask = gen_mask(kspaces[0, ..., 0], accel_factor=self.af)
-        fourier_mask = np.repeat(mask.astype(np.float), k_shape[0], axis=0)
+        fourier_mask = np.repeat(mask.astype(np.float32), k_shape[0], axis=0)
         fourier_op = FFT2(fourier_mask)
         for i, image in enumerate(images):
             kspaces[i] = fourier_op.op(image[..., 0])[..., None]
@@ -245,7 +245,7 @@ class KIKISequence(Oasis2DSequence):
         kspaces = np.empty_like(images, dtype=np.complex64)
         kspaces_masked = np.empty_like(images, dtype=np.complex64)
         mask = gen_mask(kspaces[0, ..., 0], accel_factor=self.af)
-        fourier_mask = np.repeat(mask.astype(np.float), k_shape[0], axis=0)
+        fourier_mask = np.repeat(mask.astype(np.float32), k_shape[0], axis=0)
         fourier_op = FFT2(np.array([1]))
         for i, image in enumerate(images):
             kspaces[i] = fourier_op.op(image[..., 0])[..., None]
